// src/types/seo.ts
export type SeoSettings = {
  site_name?: string;
  site_description?: string;
  site_keywords?: string;
  site_author?: string;
  site_robots?: string;
  og_title?: string;
  og_description?: string;
  og_image?: string;
  og_type?: string;
  twitter_card?: string;
  twitter_site?: string;
  twitter_creator?: string;
};